moved mock-hook-blockchain, moved mock, removed mock-hook-crypto
