+++ arwenmandos/util.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 	"math/big"
 
-	worldhook "github.com/ElrondNetwork/arwen-wasm-vm/test/mock-hook-blockchain"
+	worldhook "github.com/ElrondNetwork/arwen-wasm-vm/mock/world"
 
 	mj "github.com/ElrondNetwork/arwen-wasm-vm/test/test-util/mandos/json/model"
 	oj "github.com/ElrondNetwork/arwen-wasm-vm/test/test-util/orderedjson"
