+++ mock/context/vmHostStub.go
