+++ mock/world/worldDef.go
